+++ common/src/main/java/com/nemonotfound/nemos/inventory/sorting/mixin/ContainerScreenMixin.java
@@ -1,6 +1,8 @@
 package com.nemonotfound.nemos.inventory.sorting.mixin;
 
 import com.nemonotfound.nemos.inventory.sorting.client.ModKeyMappings;
+import com.nemonotfound.nemos.inventory.sorting.client.config.ComponentConfig;
+import com.nemonotfound.nemos.inventory.sorting.client.config.ConfigUtil;
 import com.nemonotfound.nemos.inventory.sorting.client.gui.components.AbstractSortButton;
 import com.nemonotfound.nemos.inventory.sorting.factory.*;
 import net.minecraft.client.KeyMapping;
@@ -16,12 +18,17 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
+
+import static com.nemonotfound.nemos.inventory.sorting.Constants.*;
 
 @Mixin(ContainerScreen.class)
 public abstract class ContainerScreenMixin extends AbstractContainerScreen<ChestMenu> {
 
-    @Shadow @Final private int containerRows;
+    @Shadow
+    @Final
+    private int containerRows;
     @Unique
     private final int nemosInventorySorting$containerSize = this.getMenu().getContainer().getContainerSize();
     @Unique
@@ -45,14 +52,7 @@ public abstract class ContainerScreenMixin extends AbstractContainerScreen<Chest
 
     @Unique
     private void nemosInventorySorting$initButtons() {
-        int xOffsetFirstButton = 22;
-        int xOffsetSecondButton = 40;
-        int xOffsetThirdButton = 58;
-        int xOffsetFourthButton = 76;
-        int xOffsetFifthButton = 94;
         int yOffsetInventory = 18 + (containerRows * 18);
-        int yOffsetContainer = 5;
-        int size = 11;
 
         SortAlphabeticallyButtonFactory sortAlphabeticallyButtonFactory = SortAlphabeticallyButtonFactory.getInstance();
         SortAlphabeticallyDescendingButtonFactory sortAlphabeticallyDescendingButtonFactory = SortAlphabeticallyDescendingButtonFactory.getInstance();
@@ -60,32 +60,72 @@ public abstract class ContainerScreenMixin extends AbstractContainerScreen<Chest
         MoveSameButtonFactory moveSameButtonFactory = MoveSameButtonFactory.getInstance();
         MoveAllButtonFactory moveAllButtonFactory = MoveAllButtonFactory.getInstance();
 
-        nemosInventorySorting$createContainerButton(ModKeyMappings.SORT_ALPHABETICALLY.get(), sortAlphabeticallyButtonFactory, xOffsetSecondButton, yOffsetContainer, size);
-        nemosInventorySorting$createContainerButton(ModKeyMappings.SORT_ALPHABETICALLY_DESCENDING.get(), sortAlphabeticallyDescendingButtonFactory, xOffsetFirstButton, yOffsetContainer, size);
-        nemosInventorySorting$createContainerButton(ModKeyMappings.MOVE_SAME.get(), moveSameButtonFactory, xOffsetThirdButton, yOffsetContainer, size);
-        nemosInventorySorting$createContainerButton(ModKeyMappings.MOVE_ALL.get(), moveAllButtonFactory, xOffsetFourthButton, yOffsetContainer, size);
-        nemosInventorySorting$createContainerButton(ModKeyMappings.DROP_ALL.get(), dropAllButtonFactory, xOffsetFifthButton, yOffsetContainer, size);
+        var configs = ConfigUtil.readConfigs();
 
-        nemosInventorySorting$createInventoryButton(ModKeyMappings.SORT_ALPHABETICALLY_INVENTORY.get(), sortAlphabeticallyButtonFactory, xOffsetSecondButton, yOffsetInventory, size);
-        nemosInventorySorting$createInventoryButton(ModKeyMappings.SORT_ALPHABETICALLY_DESCENDING_INVENTORY.get(), sortAlphabeticallyDescendingButtonFactory, xOffsetFirstButton, yOffsetInventory, size);
-        nemosInventorySorting$createInventoryButton(ModKeyMappings.MOVE_SAME_INVENTORY.get(), moveSameButtonFactory, xOffsetThirdButton, yOffsetInventory, size);
-        nemosInventorySorting$createInventoryButton(ModKeyMappings.MOVE_ALL_INVENTORY.get(), moveAllButtonFactory, xOffsetFourthButton, yOffsetInventory, size);
-        nemosInventorySorting$createInventoryButton(ModKeyMappings.DROP_ALL_INVENTORY.get(), dropAllButtonFactory, xOffsetFifthButton, yOffsetInventory, size);
+        nemosInventorySorting$createButtonForContainer(configs, SORT_ALPHABETICALLY_CONTAINER, ModKeyMappings.SORT_ALPHABETICALLY.get(), sortAlphabeticallyButtonFactory, X_OFFSET_SORT_ALPHABETICALLY_CONTAINER, Y_OFFSET_CONTAINER, BUTTON_SIZE, BUTTON_SIZE);
+        nemosInventorySorting$createButtonForContainer(configs, SORT_ALPHABETICALLY_DESCENDING_CONTAINER, ModKeyMappings.SORT_ALPHABETICALLY_DESCENDING.get(), sortAlphabeticallyDescendingButtonFactory, X_OFFSET_SORT_ALPHABETICALLY_DESCENDING_CONTAINER, Y_OFFSET_CONTAINER, BUTTON_SIZE, BUTTON_SIZE);
+        nemosInventorySorting$createButtonForContainer(configs, MOVE_SAME_CONTAINER, ModKeyMappings.MOVE_SAME.get(), moveSameButtonFactory, X_OFFSET_MOVE_SAME_CONTAINER, Y_OFFSET_CONTAINER, BUTTON_SIZE, BUTTON_SIZE);
+        nemosInventorySorting$createButtonForContainer(configs, MOVE_ALL_CONTAINER, ModKeyMappings.MOVE_ALL.get(), moveAllButtonFactory, X_OFFSET_MOVE_ALL_CONTAINER, Y_OFFSET_CONTAINER, BUTTON_SIZE, BUTTON_SIZE);
+        nemosInventorySorting$createButtonForContainer(configs, DROP_ALL_CONTAINER, ModKeyMappings.DROP_ALL.get(), dropAllButtonFactory, X_OFFSET_DROP_ALL_CONTAINER, Y_OFFSET_CONTAINER, BUTTON_SIZE, BUTTON_SIZE);
+
+        nemosInventorySorting$createButtonForInventory(configs, SORT_ALPHABETICALLY_CONTAINER_INVENTORY, ModKeyMappings.SORT_ALPHABETICALLY_INVENTORY.get(), sortAlphabeticallyButtonFactory, X_OFFSET_SORT_ALPHABETICALLY_CONTAINER, yOffsetInventory, BUTTON_SIZE, BUTTON_SIZE);
+        nemosInventorySorting$createButtonForInventory(configs, SORT_ALPHABETICALLY_DESCENDING_CONTAINER_INVENTORY, ModKeyMappings.SORT_ALPHABETICALLY_DESCENDING_INVENTORY.get(), sortAlphabeticallyDescendingButtonFactory, X_OFFSET_SORT_ALPHABETICALLY_DESCENDING_CONTAINER, yOffsetInventory, BUTTON_SIZE, BUTTON_SIZE);
+        nemosInventorySorting$createButtonForInventory(configs, MOVE_SAME_CONTAINER_INVENTORY, ModKeyMappings.MOVE_SAME_INVENTORY.get(), moveSameButtonFactory, X_OFFSET_MOVE_SAME_CONTAINER, yOffsetInventory, BUTTON_SIZE, BUTTON_SIZE);
+        nemosInventorySorting$createButtonForInventory(configs, MOVE_ALL_CONTAINER_INVENTORY, ModKeyMappings.MOVE_ALL_INVENTORY.get(), moveAllButtonFactory, X_OFFSET_MOVE_ALL_CONTAINER, yOffsetInventory, BUTTON_SIZE, BUTTON_SIZE);
+        nemosInventorySorting$createButtonForInventory(configs, DROP_ALL_CONTAINER_INVENTORY, ModKeyMappings.DROP_ALL_INVENTORY.get(), dropAllButtonFactory, X_OFFSET_DROP_ALL_CONTAINER, yOffsetInventory, BUTTON_SIZE, BUTTON_SIZE);
     }
 
     @Unique
-    private void nemosInventorySorting$createContainerButton(KeyMapping keyMapping, ButtonCreator buttonCreator, int xOffset, int yOffset, int size) {
-        nemosInventorySorting$createButton(keyMapping, buttonCreator, 0, nemosInventorySorting$containerSize, xOffset, yOffset, size);
+    private void nemosInventorySorting$createButtonForContainer(List<ComponentConfig> configs, String componentName, KeyMapping keyMapping, ButtonCreator buttonCreator, int defaultXOffset, int defaultYOffset, int defaultWidth, int defaultHeight) {
+        var optionalComponentConfig = ConfigUtil.getConfigs(configs, componentName);
+
+        if (optionalComponentConfig.isEmpty()) {
+            nemosInventorySorting$createContainerButton(keyMapping, buttonCreator, defaultXOffset, defaultYOffset, defaultWidth, defaultHeight);
+            return;
+        }
+
+        var config = optionalComponentConfig.get();
+
+        if (!config.isEnabled()) {
+            return;
+        }
+
+        var yOffset = config.yOffset() != null ? config.yOffset() : defaultYOffset;
+        nemosInventorySorting$createContainerButton(keyMapping, buttonCreator, config.xOffset(), yOffset, config.width(), config.height());
     }
 
     @Unique
-    private void nemosInventorySorting$createInventoryButton(KeyMapping keyMapping, ButtonCreator buttonCreator, int xOffset, int yOffset, int size) {
-        nemosInventorySorting$createButton(keyMapping, buttonCreator, nemosInventorySorting$containerSize, nemosInventorySorting$inventoryEndIndex, xOffset, yOffset, size);
+    private void nemosInventorySorting$createButtonForInventory(List<ComponentConfig> configs, String componentName, KeyMapping keyMapping, ButtonCreator buttonCreator, int defaultXOffset, int defaultYOffset, int defaultWidth, int defaultHeight) {
+        var optionalComponentConfig = ConfigUtil.getConfigs(configs, componentName);
+
+        if (optionalComponentConfig.isEmpty()) {
+            nemosInventorySorting$createInventoryButton(keyMapping, buttonCreator, defaultXOffset, defaultYOffset, defaultWidth, defaultHeight);
+            return;
+        }
+
+        var config = optionalComponentConfig.get();
+
+        if (!config.isEnabled()) {
+            return;
+        }
+
+        var yOffset = config.yOffset() != null ? config.yOffset() : defaultYOffset;
+        nemosInventorySorting$createInventoryButton(keyMapping, buttonCreator, config.xOffset(), yOffset, config.width(), config.height());
     }
 
     @Unique
-    private void nemosInventorySorting$createButton(KeyMapping keyMapping, ButtonCreator buttonCreator, int startIndex, int endIndex, int xOffset, int yOffset, int size) {
-        var sortButton = buttonCreator.createButton(startIndex, endIndex, leftPos, topPos, xOffset, yOffset, imageWidth, size, size, this);
+    private void nemosInventorySorting$createContainerButton(KeyMapping keyMapping, ButtonCreator buttonCreator, int xOffset, int yOffset, int width, int height) {
+        nemosInventorySorting$createButton(keyMapping, buttonCreator, 0, nemosInventorySorting$containerSize, xOffset, yOffset, width, height);
+    }
+
+    @Unique
+    private void nemosInventorySorting$createInventoryButton(KeyMapping keyMapping, ButtonCreator buttonCreator, int xOffset, int yOffset, int width, int height) {
+        nemosInventorySorting$createButton(keyMapping, buttonCreator, nemosInventorySorting$containerSize, nemosInventorySorting$inventoryEndIndex, xOffset, yOffset, width, height);
+    }
+
+    @Unique
+    private void nemosInventorySorting$createButton(KeyMapping keyMapping, ButtonCreator buttonCreator, int startIndex, int endIndex, int xOffset, int yOffset, int width, int height) {
+        var sortButton = buttonCreator.createButton(startIndex, endIndex, leftPos, topPos, xOffset, yOffset, imageWidth, width, height, this);
         nemosInventorySorting$keyMappingButtonMap.put(keyMapping, sortButton);
     }
 
@@ -95,7 +135,7 @@ public abstract class ContainerScreenMixin extends AbstractContainerScreen<Chest
                 .filter(entry -> entry.getKey().matches(keyCode, scanCode))
                 .findFirst();
 
-       if (keyCode == 340) {
+        if (keyCode == 340) {
             nemosInventorySorting$updateToolTips(true);
         } else {
             optionalButtonEntry.ifPresent(entry -> {