+++ Changelog.md
@@ -1,7 +1,10 @@
 # Changelog v1.8
 
+## Additions
+- Added configuration to disable, move and resize all components
+
 ## Changes
-- Search box should be usable in more modded chests now
+- Enchantments are now also included in the filter (including enchantment level)
 
 ## Fixes
 - Dimmed and highlighted slots are aligned incorrectly in shulker boxes